Moved BlockPtr and Gang out of zfs.rs
